site: Add base config & customizations
Add the base configuration for the site moving the configuration from
one single file to a folder to be able to split it in several files for
better organization.

Add a theme based on my fork because of some fixes that I sent, but they
haven't been merged yet.

Customize the header to support children menus for the main menu and the
banner image.

Add a couple of shortcodes that I plan to use in my content.