content: Add the initial content of the website
This commit add the content of the homepage, which is defined through
the languages configuration file.

The plan is to have the site in 2 languages (Catalan and English), at
least the main content, which is the one added in this commit.

Add the content pages of the initial and minimal information that the
site must contain before publishing it.

Because the theme doesn't have the i18n for Catalan, add it.