site: Make Google forms more responsive
Adjust the Google form short code to be more responsive friendly and
update the width and height arguments on the pages that use it.

This make the forms not to overflow horizontally on small screens and
the height works without showing a scroll in the iframe for big screens,
while in smaller screens the scroll show up for avoiding a vertical
overflow.