content: Move past event to a new section
We did our first event, so it isn't an upcoming event anymore.

The site didn't have a section for past events, so I created one with
the same layout than the events-next section, but changing the title
accordingly.

I moved this past event to this new section and I created an index page
for the events-next section to show a message that we don't have any
upcoming events, but we are organizing them. The new page is an alias
of the previous URL (a.k.a client-side redirect) to keep SEO happy.

I added a new entry in the main right-side menu to access the past
events section (i.e. events-past).