content: Manually disable the index of events-next section
Manually disable the index page for the events-next section. This
consist in renaming the pages and make them as draft.

In the future, I'd like to have an index page that shows the list of
upcoming events or show that there aren't events when there isn't any
future planned one and not having to disable or enabled manually.